add some feature to test
